+++ tests/integration/test_sdv.py
@@ -1,4 +1,7 @@
+from operator import attrgetter
+
 from sdv import SDV, load_demo
+from tests.integration import datasets
 
 
 def test_sdv():
@@ -72,6 +75,24 @@ def test_sdv_multiparent():
     assert set(character_families.columns) == set(tables['character_families'].columns)
 
 
+def test_sdv_multi_foreign_key():
+    """Ensure multi-foreign-key datasets are properly covered.
+
+    Multi-foreign-key datasets are those that have one table with
+    2 foreign keys to the same parent.
+    """
+    metadata, tables = datasets.load_multi_foreign_key()
+
+    sdv = SDV()
+    sdv.fit(metadata, tables)
+
+    # Sample all
+    sampled = sdv.sample()
+
+    assert set(sampled.keys()) == {'parent', 'child'}
+    assert len(sampled['parent']) == 10
+
+
 def test_integer_categoricals():
     """Ensure integer categoricals are still sampled as integers.
 
@@ -88,5 +109,6 @@ def test_integer_categoricals():
     sdv.fit(metadata, tables)
     sampled = sdv.sample()
 
+    kind = attrgetter('kind')
     for name, table in tables.items():
-        assert (sampled[name].dtypes == table.dtypes).all()
+        assert (sampled[name].dtypes.apply(kind) == table.dtypes.apply(kind)).all()