+++ tests/unit/metadata/test_table.py
@@ -1,0 +1,19 @@
+import pandas as pd
+import pytest
+
+from sdv.metadata import Table
+
+
+class TestTable:
+
+    def test__make_ids(self):
+        """Test whether regex is correctly generating expressions."""
+        metadata = {'subtype': 'string', 'regex': '[a-d]'}
+        keys = Table._make_ids(metadata, 3)
+        assert (keys == pd.Series(['a', 'b', 'c'])).all()
+
+    def test__make_ids_fail(self):
+        """Test if regex fails with more requested ids than available unique values."""
+        metadata = {'subtype': 'string', 'regex': '[a-d]'}
+        with pytest.raises(ValueError):
+            Table._make_ids(metadata, 20)